refactor: applied review to pr
